Phase transition polymer-graphene nano-composite
In the first instance, this program develops a polymer environment in which graphene nanoparticles are embedded randomly. As the environment becomes saturated, random walks interact with the environment due to a Hamiltonian potential (this is simulated with phase-type distribution functions). Finally, we calculate the average time to reach the trajectories' absorption state for each saturation index.